+++ Demo/tests/integration/test_api.py
@@ -1,8 +1,9 @@
 """
-Integration tests for API endpoints.
-These tests use FastAPI TestClient to test the full request/response cycle.
+Integration tests for all API endpoints.
+Comprehensive testing including success scenarios, error handling, and edge cases.
 """
 from unittest.mock import AsyncMock, patch
+from datetime import datetime, timezone
 
 from app.models.finding_input import FindingInput, Finding
 
@@ -197,6 +198,134 @@ class TestProcessFindingsEndpoint:
         assert response.status_code == 401
         assert "Invalid API key" in response.text
     
+    def test_process_findings_submission_before_start_time(self, client):
+        """Test submission before task start time."""
+        from app.types import Task
+        
+        # Task with future start time
+        future_task = Task(
+            taskId="test-task-123",
+            projectRepo="https://example.com/repo.git",
+            title="Test Task",
+            description="Test",
+            bounty=None,
+            status="Open",
+            startTime=str(int(datetime(2030, 1, 1, 0, 0, 0, tzinfo=timezone.utc).timestamp())),  # Future
+            deadline="1893456000",
+            selectedBranch="main",
+            selectedFiles=["contracts/Vault.sol"],
+            selectedDocs=[],
+            additionalLinks=[],
+            additionalDocs=None,
+            qaResponses=[],
+            commitSha="abc123"
+        )
+        
+        findings_data = FindingInput(
+            task_id="test-task-123",
+            findings=[
+                Finding(
+                    title="Early Submission",
+                    description="This submission is too early",
+                    severity="Medium",
+                    file_paths=["test.sol"]
+                )
+            ]
+        )
+        
+        client.mock_db.get_agent_id = AsyncMock(return_value="test-agent")
+        client.mock_db.get_task = AsyncMock(return_value=future_task)
+        
+        response = client.post(
+            "/process_findings",
+            headers={"X-API-Key": "test-key"},
+            json=findings_data.model_dump()
+        )
+        
+        assert response.status_code == 403
+        assert "Submission period has not started yet" in response.text
+    
+    def test_process_findings_submission_after_deadline(self, client):
+        """Test submission after task deadline."""
+        from app.types import Task
+        
+        # Task with past deadline
+        past_task = Task(
+            taskId="test-task-123",
+            projectRepo="https://example.com/repo.git",
+            title="Test Task",
+            description="Test",
+            bounty=None,
+            status="Open",
+            startTime="1000000000",  # Past
+            deadline=str(int(datetime(2020, 1, 1, 0, 0, 0, tzinfo=timezone.utc).timestamp())),  # Past
+            selectedBranch="main",
+            selectedFiles=["contracts/Vault.sol"],
+            selectedDocs=[],
+            additionalLinks=[],
+            additionalDocs=None,
+            qaResponses=[],
+            commitSha="abc123"
+        )
+        
+        findings_data = FindingInput(
+            task_id="test-task-123",
+            findings=[
+                Finding(
+                    title="Late Submission",
+                    description="This submission is too late",
+                    severity="Low",
+                    file_paths=["test.sol"]
+                )
+            ]
+        )
+        
+        client.mock_db.get_agent_id = AsyncMock(return_value="test-agent")
+        client.mock_db.get_task = AsyncMock(return_value=past_task)
+           
+        response = client.post(
+            "/process_findings",
+            headers={"X-API-Key": "test-key"},
+            json=findings_data.model_dump()
+        )
+        
+        assert response.status_code == 403
+        assert "Submission period has ended" in response.text
+    
+    @patch('app.main.post_submission')
+    def test_process_findings_database_error(self, mock_post_sub, sample_task, client):
+        """Test process_findings when database operations fail."""
+        # Setup mocks
+        client.mock_db.delete_agent_findings = AsyncMock(return_value=0)
+        client.mock_db.create_finding = AsyncMock(side_effect=Exception("Database error"))
+        mock_post_sub.return_value = AsyncMock()
+        
+        findings_data = FindingInput(
+            task_id="test-task-123",
+            findings=[
+                Finding(
+                    title="Database Error Test",
+                    description="This should cause a database error",
+                    severity="Medium",
+                    file_paths=["test.sol"]
+                )
+            ]
+        )
+        
+        client.mock_db.get_agent_id = AsyncMock(return_value="test-agent")
+        client.mock_db.get_task = AsyncMock(return_value=sample_task)
+        client.mock_db.delete_agent_findings = AsyncMock(return_value=0)
+        client.mock_db.create_finding = AsyncMock(side_effect=Exception("Database error"))
+            
+        response = client.post(
+            "/process_findings",
+            headers={"X-API-Key": "test-key"},
+            json=findings_data.model_dump()
+        )
+        
+        assert response.status_code == 500
+        assert "Error processing findings" in response.text
+    
     def test_process_findings_missing_api_key(self, client):
         """Test findings submission without API key."""
         findings_data = FindingInput(
@@ -323,6 +452,94 @@ class TestBackgroundProcessingEndpoint:
         
         assert response.status_code == 400
         assert "Invalid test task ID" in response.text
+    
+    def test_background_processing_invalid_agent_authentication(self, client):
+        """Test background processing when agent authentication fails."""
+        findings_data = FindingInput(
+            task_id="TESTTASK",
+            findings=[
+                Finding(
+                    title="Test Finding",
+                    description="Test description",
+                    severity="High",
+                    file_paths=["test.sol"]
+                )
+            ]
+        )
+        
+        # Mock get_agent_id to raise ValueError (invalid API key)
+        client.mock_db.get_agent_id = AsyncMock(side_effect=ValueError("Invalid API key"))
+            
+        response = client.post(
+            "/test/process_findings",
+            headers={"X-API-Key": "invalid-key"},
+            json=findings_data.model_dump()
+        )
+        
+        assert response.status_code == 401
+        assert "Invalid API key" in response.text
+    
+    @patch('app.main.config')
+    def test_background_processing_max_findings_exceeded(self, mock_config, client):
+        """Test background processing with too many findings."""
+        # Set low limit for this test
+        mock_config.max_findings_per_submission = 1
+        
+        # Create more findings than allowed
+        findings = []
+        for i in range(2):
+            findings.append(Finding(
+                title=f"Test Finding {i+1}",
+                description=f"Description {i+1}",
+                severity="Medium",
+                file_paths=["test.sol"]
+            ))
+        
+        findings_data = FindingInput(
+            task_id="TESTTASK",
+            findings=findings
+        )
+        
+        client.mock_db.get_agent_id = AsyncMock(return_value="test-agent")
+            
+        response = client.post(
+            "/test/process_findings",
+            headers={"X-API-Key": "test-key"},
+            json=findings_data.model_dump()
+        )
+        
+        assert response.status_code == 400
+        assert "Maximum allowed: 1" in response.text
+    
+    @patch('app.main.post_submission')
+    def test_background_processing_database_error(self, mock_post_sub, client):
+        """Test background processing when database operations fail."""
+        mock_post_sub.return_value = AsyncMock()
+        
+        findings_data = FindingInput(
+            task_id="TESTTASK",
+            findings=[
+                Finding(
+                    title="Database Error Test",
+                    description="This should cause a database error",
+                    severity="Medium",
+                    file_paths=["test.sol"]
+                )
+            ]
+        )
+        
+        client.mock_db.get_agent_id = AsyncMock(return_value="test-agent")
+        # Setup mongodb mock to raise exception
+        client.mock_db.create_finding = AsyncMock(side_effect=Exception("Database error"))
+            
+        response = client.post(
+            "/test/process_findings",
+            headers={"X-API-Key": "test-key"},
+            json=findings_data.model_dump()
+        )
+        
+        assert response.status_code == 500
+        assert "Error in test processing" in response.text
 
 
 class TestTaskFindingsEndpoint:
@@ -356,3 +573,141 @@ class TestTaskFindingsEndpoint:
         result = response.json()
         
         assert result == []
+    
+    @patch('app.main.config')
+    def test_get_task_findings_invalid_api_key(self, mock_config, client):
+        """Test getting task findings with invalid API key."""
+        mock_config.backend_api_key = "correct-key"
+        
+        response = client.get(
+            "/tasks/test-task/findings", 
+            headers={"X-API-Key": "wrong-key"}
+        )
+        
+        assert response.status_code == 401
+        assert "Invalid API key" in response.text
+    
+    @patch('app.main.config')
+    def test_get_task_findings_database_error(self, mock_config, client):
+        """Test getting task findings when database error occurs."""
+        mock_config.backend_api_key = "test-key"
+        client.mock_db.get_findings = AsyncMock(side_effect=Exception("Database connection failed"))
+        
+        response = client.get(
+            "/tasks/test-task/findings",
+            headers={"X-API-Key": "test-key"}
+        )
+        
+        assert response.status_code == 500
+        assert "Error retrieving findings" in response.text
+
+
+class TestTriggerTaskProcessingEndpoint:
+    """Test the /tasks/{task_id}/process endpoint."""
+    
+    @patch('app.main.config')
+    def test_trigger_task_processing_invalid_api_key(self, mock_config, client):
+        """Test trigger processing with invalid API key."""
+        mock_config.backend_api_key = "correct-key"
+
+        response = client.post(
+            "/tasks/test-task/process",
+            headers={"X-API-Key": "wrong-key"}
+        )
+            
+        assert response.status_code == 401
+        assert "Invalid API key" in response.text
+    
+    def test_trigger_task_processing_missing_api_key(self, client):
+        """Test trigger processing without API key header."""
+        response = client.post("/tasks/test-task/process")
+        
+        assert response.status_code == 422  # Missing required header
+    
+    @patch('app.main.mongodb')
+    @patch('app.main.config')
+    def test_trigger_task_processing_already_processed(self, mock_config, mock_mongodb, client):
+        """Test triggering processing for already processed task."""
+        mock_config.backend_api_key = "test-key"
+        mock_mongodb.get_metadata = AsyncMock(return_value={
+            "processed_at": "2025-01-01T00:00:00Z",
+            "scheduled_processing": True
+        })
+        
+        response = client.post(
+            "/tasks/test-task/process",
+            headers={"X-API-Key": "test-key"}
+        )
+        
+        assert response.status_code == 200
+        result = response.json()
+        assert result["status"] == "already_processed"
+        assert result["task_id"] == "test-task"
+    
+    @patch('app.main.mongodb') 
+    @patch('app.main.config')
+    def test_trigger_task_processing_no_pending_findings(self, mock_config, mock_mongodb, client):
+        """Test triggering processing when no pending findings exist."""
+        mock_config.backend_api_key = "test-key"
+        mock_mongodb.get_metadata = AsyncMock(return_value=None)
+        mock_mongodb.get_findings = AsyncMock(return_value=[])
+        
+        response = client.post(
+            "/tasks/test-task/process", 
+            headers={"X-API-Key": "test-key"}
+        )
+        
+        assert response.status_code == 200
+        result = response.json()
+        assert result["status"] == "no_pending_findings" 
+        assert result["task_id"] == "test-task"
+        assert result["total_findings"] == 0
+
+
+class TestPostTaskFindingsEndpoint:
+    """Test the /tasks/{task_id}/post endpoint."""
+    
+    @patch('app.main.config')
+    def test_post_task_findings_invalid_api_key(self, mock_config, client):
+        """Test posting findings with invalid API key."""
+        mock_config.backend_api_key = "correct-key"
+
+        response = client.post(
+            "/tasks/test-task/post",
+            headers={"X-API-Key": "wrong-key"}
+        )
+            
+        assert response.status_code == 401
+        assert "Invalid API key" in response.text
+    
+    @patch('app.main.config')
+    def test_post_task_findings_no_backend_endpoint(self, mock_config, client):
+        """Test posting findings when backend endpoint not configured."""
+        mock_config.backend_api_key = "test-key" 
+        mock_config.backend_findings_endpoint = None
+        
+        response = client.post(
+            "/tasks/test-task/post",
+            headers={"X-API-Key": "test-key"}
+        )
+        
+        assert response.status_code == 503
+        assert "Backend findings endpoint not configured" in response.text
+    
+    @patch('app.main.config') 
+    def test_post_task_findings_no_findings(self, mock_config, client):
+        """Test posting findings when no findings exist for task."""
+        mock_config.backend_api_key = "test-key"
+        mock_config.backend_findings_endpoint = "http://test.com/findings"
+        client.mock_db.get_findings = AsyncMock(return_value=[])
+        
+        response = client.post(
+            "/tasks/test-task/post",
+            headers={"X-API-Key": "test-key"}
+        )
+        
+        assert response.status_code == 200
+        result = response.json()
+        assert result["status"] == "no_findings"
+        assert result["task_id"] == "test-task"
+        assert result["total_findings"] == 0
