+++ Demo/tests/unit/test_cache.py
@@ -146,48 +146,3 @@ class TestFetchTaskData:
             
             # Verify download was called (cache miss)
             mock_download.assert_called_once()
-
-
-@pytest.mark.asyncio
-class TestScheduleSubmittedTasks:
-    """Unit tests for scheduling functionality."""
-
-    @patch('app.main.schedule_task_processing')
-    @patch('app.main.mongodb')
-    async def test_schedule_submitted_tasks_no_database_tasks(self, mock_mongodb, mock_schedule):
-        """Test scheduling when no tasks are in database."""
-        from app import main as app_main
-        
-        mock_mongodb.get_submitted_tasks = AsyncMock(return_value=[])
-        mock_mongodb.schedule_task_processing = AsyncMock()
-            
-        await app_main.schedule_submitted_tasks()
-        
-        mock_mongodb.get_submitted_tasks.assert_called_once()
-        mock_schedule.assert_not_called()
-
-    @patch('app.main.mongodb')
-    async def test_schedule_submitted_tasks_database_error(self, mock_mongodb):
-        """Test scheduling when database error occurs."""
-        from app import main as app_main
-        
-        mock_mongodb.get_submitted_tasks = AsyncMock(side_effect=Exception("Database error"))
-
-        # Should not raise exception
-        await app_main.schedule_submitted_tasks()
-
-    @patch('app.main.schedule_task_processing')
-    @patch('app.main.mongodb')
-    async def test_testtask_not_scheduled(self, mock_mongodb, mock_schedule):
-        """Test that TESTTASK doesn't trigger job scheduling."""
-        from app import main as app_main
-        from tests.conftest import create_sample_task
-        
-        tasks = [create_sample_task(task_id="TESTTASK")]
-        
-        mock_mongodb.get_submitted_tasks = AsyncMock(return_value=tasks)
-            
-        await app_main.schedule_submitted_tasks()
-            
-        # Should not schedule TESTTASK
-        mock_schedule.assert_not_called()
